Bonus number 테스트 분리 , 메소드명 수정
