+++ src/test/java/secondsteptest/LottoMakerTest.java
@@ -25,8 +25,8 @@ public class LottoMakerTest {
     }
 
     @Test
-    @DisplayName("로또 번호가 중복으로 안 만들어지는지 확인")
-    void create_not_duplication_lotto_number() {
+    @DisplayName("리스트에 값을 넣은 후 shuffle 을 사용하면 다른 리스트가 되는지 확인한다.")
+    void list_collections_shuffle_test() {
         List<Integer> actual = new ArrayList<>();
         List<Integer> excepted = new ArrayList<>();
         for (int i = 0; i < 7; i++) {
