+++ src/test/java/secondsteptest/ResultViewTest.java
@@ -43,14 +43,5 @@ public class ResultViewTest {
                 .hasMessage("지난 주 당첨 번호는 6개 입니다.");
     }
 
-    @ParameterizedTest(name = "보너스 번호는 1에서 45 사이의 숫자가 아니면 예외가 발생한다 :[{index}] : [{arguments}]")
-    @DisplayName("보너스 번호는 1에서 45 사이의 숫자가 아니면 예외가 발생한다")
-    @ValueSource(ints = {
-            0, 46
-    })
-    void throw_exception_bonus_number_is_one_between_forty_five(int bonusNumbers) {
-        assertThatThrownBy(() -> resultView.bonusLottoNumberValidation(bonusNumbers))
-                .isInstanceOf(LottoNumberException.class)
-                .hasMessage("입력값: " + bonusNumbers + " 번호는 1에서 45 사이의 숫자가 아닙니다");
-    }
+
 }
